Add more complete colour properties for use in VPP
Adds colour_primaries, transfer_characteristics and matrix_coefficients
as defined by ISO/IEC 23001-8 and H.273 (and, in future, ISO/IEC 23091-2)
to the colour properties.  This allows support for HDR standards which are
not currently expressible (for example, PQ and HLG transfer functions).
These new fields are only used in VPP if the corresponding color_standard
value is set to VAColorStandardExplicit.

Also document the meaning of existing VAProcColorStandardValues in terms
of the new properties.

Change-Id: Ic6f6d506130396d09da31acc2a4d6c492a4fb390
Signed-off-by: Mark Thompson <sw@jkqxz.net>