travis: Add a test that code files don't have the exec bit set
C source code and header files shouldn't be executable.
This is a problem that already occurred pretty often in the past and was
e.g. fixed in 5d8c5ae, be95415, and 8352ab9. Therefore it seems like a
good idea to add a CI check to prevent this issue from reappearing
(probably mostly accidents due to wrong Git configurations, etc.).

Signed-off-by: Michael Weiss <dev.primeos@gmail.com>